ENG-107: rename header-nav to header
